Rework and cleanup
